requirements.txt
fastapi[all]==0.115.6
instructor==1.5.2
jiter==0.5.0
litellm==1.59.9
joblib==1.4.2
json_repair==0.35.0
langsmith==0.3.3
json5==0.9.28
kombu==5.4.2
uvicorn==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
gunicorn==23.0.0
python-dotenv==1.0.1
postgres==4.0
psycopg2-binary==2.9.10
neo4j==5.27.0
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
tqdm==4.67.1
grep-ast==0.4.1
pygments==2.18.0
networkx==3.4.2
blar-graph==1.1.6
openai==1.60.2
uuid6==2024.7.10
aiohttp==3.11.9
langchain==0.3.16
langchain-anthropic==0.3.4
langchain-deepseek-official==0.1.0
langchain-community==0.3.16
langchain-core==0.3.32
langchain-openai==0.3.2
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langchain-postgres==0.0.12
langgraph==0.2.58
firebase-admin==6.6.0
portkey-ai==1.9.7
gitPython==3.1.43
PyGithub==2.5.0
google-cloud-secret-manager==2.21.1
pre-commit==4.0.1
isort==5.13.2
black==24.10.0
ruff==0.8.1
pylint==3.3.2
bandit==1.8.0
aiofiles==24.1.0
scikit-learn==1.5.2
requests==2.32.3
resend==2.4.0
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.3.1
crewai==0.95.0
nltk==3.9.1
celery==5.4.0
redis==5.2.0
flower==2.0.1
chardet==5.2.0
sentry-sdk[fastapi]==2.20.0
posthog==3.7.4
newrelic==9.0.0
tiktoken==0.7.0
agentops==0.3.26
pydantic[email]==2.10.3
firecrawl-py==1.11.1
streamlit
google-generativeai
plotly
streamlit-mermaid
zaproxy
